backend/api/requirements.txt
Django==1.11.5
djangorestframework
django-cors-headers